misc: Split graphql definitions into two schemas
