--- app/graphql/lago_api_schema.rb
@@ -1,51 +1,0 @@
-# frozen_string_literal: true
-
-class LagoApiSchema < GraphQL::Schema
-  mutation(Types::MutationType)
-  query(Types::QueryType)
-
-  # For batch-loading (see https://graphql-ruby.org/dataloader/overview.html)
-  use GraphQL::Dataloader
-
-  max_depth 15
-  max_complexity 350
-
-  # GraphQL-Ruby calls this when something goes wrong while running a query:
-
-  # Union and Interface Resolution
-  def self.resolve_type(_abstract_type, _obj, _ctx)
-    # TODO: Implement this method
-    # to return the correct GraphQL object type for `obj`
-    raise(GraphQL::RequiredImplementationMissingError)
-  end
-
-  # Relay-style Object Identification:
-
-  # Return a string UUID for `object`
-  def self.id_from_object(object, type_definition, _query_ctx)
-    # For example, use Rails' GlobalID library (https://github.com/rails/globalid):
-    object_id = object.to_global_id.to_s
-    # Remove this redundant prefix to make IDs shorter:
-    object_id = object_id.sub("gid://#{GlobalID.app}/", '')
-    encoded_id = Base64.urlsafe_encode64(object_id)
-    # Remove the "=" padding
-    encoded_id = encoded_id.sub(/=+/, '')
-    # Add a type hint
-    type_hint = type_definition.graphql_name.first
-    "#{type_hint}_#{encoded_id}"
-  end
-
-  # Given a string UUID, find the object
-  def self.object_from_id(encoded_id_with_hint, _query_ctx)
-    # For example, use Rails' GlobalID library (https://github.com/rails/globalid):
-    # Split off the type hint
-    _type_hint, encoded_id = encoded_id_with_hint.split('_', 2)
-    # Decode the ID
-    id = Base64.urlsafe_decode64(encoded_id)
-    # Rebuild it for Rails then find the object:
-    full_global_id = "gid://#{GlobalID.app}/#{id}"
-    GlobalID::Locator.locate(full_global_id)
-  end
-
-  default_max_page_size 25
-end
