+++ app/graphql/types/customer_portal/mutation_type.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Types
+  module CustomerPortal
+    class MutationType < Types::BaseObject
+      field :download_customer_portal_invoice, mutation: Mutations::CustomerPortal::DownloadInvoice
+    end
+  end
+end
