+++ spec/graphql/concerns/authenticable_api_user_spec.rb
@@ -24,7 +24,7 @@ module AuthenticableApiUserSpec
     field :renameThing, mutation: RenameThingMutation
   end
 
-  class TestApiSchema < LagoApiSchema
+  class TestApiSchema < Schemas::ApiSchema
     mutation(ThingsMutationType)
   end
 end
