+++ spec/graphql/concerns/authenticable_customer_portal_user_spec.rb
@@ -21,7 +21,7 @@ module AuthenticableCustomerPortalUserSpec
     field :thing, resolver: ThingResolver
   end
 
-  class TestApiSchema < LagoApiSchema
+  class TestApiSchema < Schemas::CustomerPortalSchema
     query(ThingsQueryType)
   end
 end
