+++ spec/graphql/concerns/can_require_permissions_spec.rb
@@ -26,7 +26,7 @@ module CanRequirePermissionsSpec
     field :renameThing, mutation: RenameThingMutation
   end
 
-  class TestApiSchema < LagoApiSchema
+  class TestApiSchema < Schemas::ApiSchema
     mutation(ThingsMutationType)
   end
 end
